# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Lazy AI Coder - GitLab integration, LLM tools, and MCP server\n\nFeatures:\n- LLM chat and code analysis\n- GitLab integration (read files, review MRs)\n- MCP (Model Context Protocol) server\n- PlantUML diagram generation\n- Session management with memory",
        "title": "Lazy AI Coder API",
        "contact": {
            "name": "API Support",
            "email": "walter.fan@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/prompts": {
            "get": {
                "description": "Get list of prompts with filtering and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prompts"
                ],
                "summary": "List prompts",
                "parameters": [
                    {
                        "type": "string",
                        "default": "all",
                        "description": "Scope: all, personal, shared, templates",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search in name and description",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by tags (comma-separated for OR logic)",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "created_at",
                        "description": "Sort by: created_at, updated_at, name",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (1-based, overrides offset)",
                        "name": "pageNumber",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 200,
                        "description": "Page size (overrides limit)",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 200,
                        "description": "Limit results (deprecated, use pageSize)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination (deprecated, use pageNumber)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Prompt"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new prompt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prompts"
                ],
                "summary": "Create prompt",
                "parameters": [
                    {
                        "description": "Prompt data",
                        "name": "prompt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreatePromptRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Prompt"
                        }
                    }
                }
            }
        },
        "/api/v1/prompts/{id}": {
            "get": {
                "description": "Get a specific prompt by ID (UUID) or Name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prompts"
                ],
                "summary": "Get prompt by ID or Name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Prompt ID or Name",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Prompt"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing prompt by ID (UUID) or Name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prompts"
                ],
                "summary": "Update prompt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Prompt ID or Name",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Prompt data",
                        "name": "prompt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdatePromptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Prompt"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a prompt (soft delete) by ID (UUID) or Name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prompts"
                ],
                "summary": "Delete prompt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Prompt ID or Name",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreatePromptRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "arguments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PromptArgument"
                    }
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "scope": {
                    "description": "\"personal\" or \"shared\"",
                    "type": "string"
                },
                "system_prompt": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_prompt": {
                    "type": "string"
                }
            }
        },
...
        "handlers.UpdatePromptRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "arguments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PromptArgument"
                    }
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "system_prompt": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_prompt": {
                    "type": "string"
                }
            }
        },
...
        "models.Prompt": {
            "type": "object",
            "properties": {
                "created_by": {
                    "type": "string"
                },
                "created_time": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "realm_id": {
                    "description": "NULL for global templates",
                    "type": "string"
                },
                "system_prompt": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "title": {
                    "description": "Human-readable display name",
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "updated_time": {
                    "type": "string"
                },
                "user_id": {
                    "description": "NULL for global templates",
                    "type": "string"
                },
                "user_prompt": {
                    "type": "string"
                }
            }
        },
        "models.PromptArgument": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}